clustersConfig: add "test_only" flag to ClustersConfig() to only parse YAML for testing
This constructor does several things. In particular, the latter parts of
it also call "virsh" or run other local command. That makes unit testing
the class(and the parsing of the YAML) hard. Point in case: there
currently is no unit test.

Add a flag to abort half way through, so that we can at least test the
first part. Unit tests will be added next and use this.

Later, once we have tests and cleanup the code some more, the parsing of
the YAML will be separate from additional setup steps, and the test can
then only test the parsing. Then the flag will be dropped again.