clustersConfig: move IP range calculation to ClusterConfig
That way, it's going to be available for every ClusterConfig, not just
the first cc.main_config.clusters[0].

Also, _parse_ip_range() is now clearly without side effect (except the
"get_last_ip" callback that we pass it). This makes it unit testable and
easier to reason about.

Note that previously, we used `self.all_nodes`, which honors the
self.worker_range. Now, instead we iterate over all workers. That seems
better, because if you only want to configure certain workers, we still
should get the configuration that applies when using them all.