clustersConfig: drop detection of last_ip for the IP range
The IP addresses are commonly configured in the cluster YAML. Also, the
"ip_range" can be configured there.

It seems odd, to fetch a range from virsh, and let that affect the
configuration.

Instead, the configuration YAML (and its potential defaults) should
determine the range. If virsh disagrees, we must reconfigure libvirt and
not the other way around.

Compare to "external_port: auto", where we also pick up external
configuration from the system. But for the IP range, we have the node IP
addresses in the YAML too, and the detected range must match with that
configuration. Detecting the range while the IP addresses in the YAML
depend on it is problematic.

Also, because the very first time when you run CDA deploy on a host,
this will not pick up anything, because virsh is not configured.

The benefit is, that this removes an external source for teh
configuration and it depends only on the YAML.