clusterDeployer: fix is-bf check in ClusterDeployer
self.is_bf was assigned a generator. The later check `if not
self.is_bf:` would never evaluate to true and "iso_path = os.getcwd()"
was never executed.

Fix that. Note that this changes (fixes?) behavior for the common case,
where we don't have "bf" nodes.

Also, a generator can only be iterated once, hence the "if not all(self.is_bf):"
check also never hit. But that was not very important, because that was
only a consistency check of the configuration.