clusterNode: add __slots__ to ClusterNode,ClusterHost classes
The benefit is that mypy will warn if we try to access any attributes
that are not explicitly listed. This helps with seeing the attributes of
a class.

Also, don't initialize the variables in the type annotation of the
class. As this is not a dataclass, this results in a class attribute.
Later uses of that attribute then may assign a per-instance value, so
this did not necessarily have bad consequences. It is still confusing to
mix class and instance attributes.