extraConfigDpu: place installed go binary in /usr/bin so that it works
While we unpack the go binary to "/usr/local/go", it seems bad to rely
on updating the PATH to the non-standard location "/usr/local/go/bin".

In particular, the setup script of CDA may not install go already (hence
the need for this ensure_go_installed()). That means, the first time you
run the DPU extra config, it will fail because the binary is not in the
path.

Instead, symlink the executable files to /usr/bin/{go,gofmt}.

Also, drop the localhost specific installation procedure. Unclear why
this is there. In particular, it is really ugly to update the
environment of a running process, don't do that. What we would want, is
to setup the PATH during host.run(). Granted, that doesn't work with
remote hosts, so ... anyway. This is unnecessary, drop it.

Also, use "curl" instead of "wget", because that is usually installed on
RHEL by default, while wget is not. It's not important whether we use
"curl" or "wget", but CDA should consistently use one or the other.