extraConfigDpu: fix using repo_dir argument in update_dockerfiles_with_ose_images()
The repo_dir is configurable in the cluster YAML configuration, see
"ExtraConfigArgs.dpu_operator_path". Using the hard coded default value
REPO_DIR is not right, instead require the caller to pass it in.

Note that there are no callers of update_dockerfiles_with_ose_images(),
so maybe this function should be just deleted. For now, that is not
done, because maybe there was a plan for this function.