extraConfigDpu: Don't start vsp "manually"
With the changes in https://github.com/openshift/dpu-operator/pull/128
we not longer need to start the vsp, this will be handled by the dpu
operator.

We will not remove the "Start" functionality, since the option to deploy
the vsp manually outside the dpu operator is maintained / will be
required for non-Intel vsps.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>