clusterInfo: add location for gspread sheet credentials and fix search
The location "~/.config/gspread/credentials.json" is what is suggested at [1].

It's nicer to use on a developer machine. I don't want to spam the top
level of my home directory, but I also don't want to place it in the
current directory (where it's subject to $PWD and possibly gets lost
during a git-clean).

Also, move the default locations to a separate function. It will be
useful later.

[1] https://docs.gspread.org/en/latest/oauth2.html#for-bots-using-service-account