dpuVendor: Build a multiarch ipu manifest
The dpu-operator by default expects a multiarch vsp image.

In the past, this code just worked, because we would build and push the
image immediately before deploying the operator on the respective
architecture, so the most recent image would be the proper arch.

This is flimsy, and can cause errors later on if we need to pull the
image again. Instead, as a workaround while waiting for multiarch IPU
vsp build support, build a manifest of the images during the second
phase of deployment (iso-cluster dpu).

This code makes a (potentially bad) assumption that the host side
deployment has already run, and built/pushed the x86_64 vsp.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>