host.py: fix subprocess bug
There is an edge case we can hit when reading from a subprocess that
will cause the read operation to hang indefinitely, which is exposed by
https://github.com/openshift/dpu-operator/pull/181/commits/c517b55f71b0dce4d2234dc5e64216305ee20c1d.

We can reach a deadlock when reading from both of these buffers
independently with a significant amount of output coming through.

It makes more sense anyways to just read from proc.communicate() than
manually reading from stdout and stderr.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>