extraConfigDpu: remove idpf netdev restart
In theory we should no longer need to manually cold boot the servers to
ensure idpf netdevs are available.

Currently there is a bug in MeV 1.8 that causes the host to reboot any
time the IMC reboots, which masks this issue from occurring.

If this masking is removed, we want to see failure here to signify that
lifecycle management of the IPU is broken

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>