libvirt: Restart libvirt services only when needed
The switch to modular from monolithic needs to happen only
once per host, disable the monolothic only if it's running.
the same applies for the modular services, there is no need
to restart them every time, just start them if they are not
running.

The restart method now takes an argument which service to
restart. This is useful especially for the virtual bridge
configuration, we don't have to restart every single service
to update the network configuration.

Signed-off-by: Ales Musil <amusil@redhat.com>