Run port setup script forever
Being clever and running the set up script for only a short ammount of
time after determining the ACC is up does not work, because when we
first install the ACC, the ACC can recieve ICMP packets during iso boot.

Instead, just run it forever.

We could do this with a cronjob, however the IMC does not have crontab /
cron functionality available by default.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>