Add workaround for flaky acc first boot
Sometimes after the first boot, things just don't work.

If we don't run the connectivity script constantly, we run the risk of
not running it in the window we need to.

If we do run the connectivity script constantly, we seem to be able to
break the node policy rules, such that there is no connectivity on any
port no matter what we do.

Instead, add a more moderate script to run ipu_port1_setup.sh, and just
reboot the IMC if things have taken too long. After the first reboot,
behavior of the ACC seems much more predictable

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>