Ensure connectivity after cold boot
We can't just cold boot and return, or else subsequent actions that
expect the dpu to be up and running might fail.

We need to make sure if we cold boot, we wait for the ACC to come back
up and validate that the openshift components are still running as
expected.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>