Cold boot host before installing dpu operator
We can just do the cold boot prior to installing the dpu operator.

Ideally, it shouldn't matter whether the ACC reboots after the
components have already come up.

However in practice, this might not work. We should add coverage for
this case in the tests, not in the workaround added to the deployment
scripts.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>