clusterNode: fix post_boot() abstraction
Previously, we have the base class ClusterNode with

  def post_boot(self, *args: Any, **kwargs: Any) -> bool

while we have subclasses implement either

  def post_boot(self) -> bool

or

  def post_boot(self, desired_ip_range: tuple[str, str]) -> bool

This is wrong. The abstract method must have an API that is suitable with
all implementations. Or at least, all implementstion must handle in some
way -- even failing -- when being alled in the various ways that the API
offers.

Maybe we could see that those different post_boot() implementations are
not actually the same, and we should have different abstract methods for
different purposes and with different names. However, that is not done
here. Instead, just ensure that all methods follow the same API.