extraConfigDpu: revert p4 pod to use container
We can't deploy the p4 component as a pod in OCP currently because of
the way the p4 pod binds the grpc service to localhost. With the current
implementation, the vsp cannot access the service at the default address
it uses (0.0.0.0:9559).

As a workaround for now, revert to using the podman container method.
Leaving the previous implementation to switch back to when this issue
has been addressed

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>