isoCluster: pass "--default-extra-packages" to marvell-tools' "pxeboot.py"
The current rhel-9 kernel lacks patches for the Marvell DPU (on the DPU
side). This will be fixed eventually, but for now, we must install
another kernel.

Marvell-tools has a command line argument to install RPMs of a suitable
kernel.  Use it.

Note that the option is just called "--default-extra-packages". In the
future, when the workaround is no longer necessary, then "marvell-tools"
will decided that it no longer needs to install any extra packages.  In
CDA, the option will always be passed to the pxeboot command.

If CDA always passes this option to pxeboot command, why does it even
exist? Because the pxeboot command can be used without CDA, and because
in that use case it's deemed better to not install the packages without
being asked to.