extraConfigDpu: rebuild vsp images based on config
If the user specifies "rebuild dpu operator images" we should also
rebuild the vsp images, since the vsp is essentially a dpu operator
image.

If the user deploys the ACC side before the host side, the manifest will
now be available for the host side to no longer have to rebuild the
images.

Note, we cannot have this behavior be completely symetrical since the
host side does not have trivial access to an arm server to build the
vsp.

The proper fix for this will be buildah multiarch support for the ipu
plugin.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>