Add preconfig step ExtraConfigMevFwUp
Provide a pre-config to put the MeV firmware into a good state.

Reflashing the firmware takes a very long time. By default, it will not
do this if the firmware is already on the desired version.

However, if we specify to force, we will reflash regardless of the
current state.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>