clusterInfo: fix load_cluster_info() to require full match
When specifying both "--host" and "--cluster" options, it seems to make
the most sense that both selectors must match. Fix that.

Now `python clusterInfo.py host -H hostname -c clustername` works as one
would(?) expect. It returns a result only if both the "-H" and "-c"
option matches, and if only a single matching result is found. Otherwise
an error is raised:

  No cluster info for host 'hostname' and for cluster  matching 'clustername' found. Select a host with "--host" or "--cluster" option?

This implementation no longer uses a dictionary lookup, it always
iterates over all entries. But since there is only a handful of entries,
the performance impact is irrelevant (especially compared to the time we
spend to fetch a document from the web).