marvell: hard code the BMC user name "core"
We want that dpu-operator's "hack/cluster-configs/config-dpu.yaml" works
with IPU and Marvell DPU.

However, the BMC user for IPU is usually "root", while for Marvell DPU
we expect this to be a CoreOS node and user "core".

For now, we don't have a way to template the bmc user name in the
cluster config (e.g. via Jinja2 templates from the lab spreadsheet).
Maybe one day that will be added.

So, for now, we will delete the "bmc.user" config from
"hack/cluster-configs/config-dpu.yaml". Note that the BMC() class still
uses "root" as default, so in code, the bmc.user is "root". We must
ignore the setting entirely and hard code it to "core".