deployment: Fail installation when one of the masters fail
The master cannot be easily retried as workers, fail the whole
installation if one of the masters fail to install. This is needed
instead of the callback trick. The callback trick will fail, however
the futures will keep running and there isn't any way to stop them
so the whole installation gets stuck anyway.

Signed-off-by: Ales Musil <amusil@redhat.com>