clusterInfo: add "secondary_network_port"
The spreadsheet can now have a value "secondary" in the "Provision host"
column. I already updated the spreadsheet already to contain this entry.

In that case, the "Ports" column contains the "ClusterInfo.secondary_network_port".

This is also accessible in CDA cluster config via Jinja2 template like

    network_api_port: {{ secondary_network_port() }}

We need this for the Marvell DPU clusters.

Note that with cluster-kind "iso" deployments, the "network_api_port" in
the cluster configuration has a different (but useful) meaning than from
"openshift" kinds. Compare to openshift clusters, where we set this to
`network_api_port: {{ api_network() }}`.

On all IPU clusters, this value is currently "eno12409". On all Marvell
DPU clusters, it is currently "eno2".

As we want that a single cluster configuration works for our IPU and
Marvell DPU clusters, and as this port differs between clusters, we must
make it configurable. This is what the patch does.