marvell: redirect output of pxeboot also to /tmp file
cluster-deployment-automation by default runs with only INFO logging
level which doesn't show what the tool does. Also, even with DEBUG logging is the output of `rsh.run()` hard to follow in this case.

To help with that, redirect the output of the pxeboot command also to
"/tmp/pxeboot-log-*" file on the host, so that we can look it up there
too.

Granted, after reboot this log in /tmp is gone. And also, with `make
e2e-test` we usually will afterwards reinstall the worker node and
reboot the machine. Thus, usually this log is lost soon after.

It can be still useful to collect it and there are situations where you
can still get this log (even it it's not persisted).