extraConfigHostRegistry.py: Make prep_auth reusable for private image mode containers
The old behavior is that we assumed the CI only needs access to the CI
registry in its function, so we hard coded that registry.
The behavior now is that prep_auth can be called multiple times to add
any new auth to access new container registry on the host.

This is useful for to be accessed by pre and post steps in our
configuration