extraConfigIsoBuilder.py: Add the kickstart generator for Image Mode
This allows us to have the same kickstart behavior as rhel for edge.
A cool bootc feature is that we can move some post boot things to the base image it
self instead of in the kickstart. This means that we can have the same
behavior in every deployment (base image) vs individual deployments (kickstart iso build time)

We also consume rhc credentials in the Image Mode kickstart generator