Revert "marvell: honor CDA_MARVELL_TOOLS_EXTRA_ARGS variable for pxeboot"
There is desire to make the tool not configurable.  At least, not
configurability via environment variables.

This option was added with the specific intent to set
"--octep-cp-agent-service-disable", while working on dpu-operator PR 303
(to allow for some configurations in CI to use/not-use the option).
This immidiate use is no longer present. So the patch will be reverted.

This reverts commit 023508b832a2c5f985ff0f4ec14c8a115eeb16e1.