bmc: cleanup hostname handling for BMC
We must be clear whether this is a "BMC.url" or a "BMC.bmc_host". That
is, is it a URL or just a hostname for which we (at certain places) need
to append a "https://" scheme. Otherwise, we get errors like

    requests.exceptions.MissingSchema: Invalid URL '10.26.16.95': No scheme supplied. Perhaps you meant https://10.26.16.95?

Rename the "url" attribute to "bmc_host" to make this clear. Also, the
various callers will instead use the "base_url" property, as
appropriate.

See-also: https://jenkins-csb-nst-net-hw-ci.dno.corp.redhat.com/job/99_E2E_Marvell_DPU_Deploy/593/console

Fixes: f857f85a82df ('Remove redundant Dell specific information from BMC class')