extraConfigMicroshift: increase number of retries for installing microshift
This command still fails, see [1].

Increase the number of retries. As there is a hard-coded sleep interval
of 5 seconds, increasing to 60 means to retry for approximately 5
minutes. There really is no harm hear.

Either the error is temporary, in which case retrying long enough to
workaround is good. Or it is permanant, in which case we will fail
slightly later after retrying for a few minutes.

[1] https://jenkins-csb-nst-net-hw-ci.dno.corp.redhat.com/view/DPU%20Testing/job/99_E2E_Marvell_DPU_Deploy/752/console