Revert "Handle errors when assisted installer goes down during iso prep"
It is unclear what the purpose of this was. The git history does not
make that clear. There was already a retry loop, bound to 15 minutes.
Restore to only rely on the retry loop.

What happens otherise is that the first pull of the ISO fails, then the
health check fails too, aborting installation and no retry. Also, on the
next run the same thing happens again, and the cluster does not recover.
If we would actually continue trying to download the ISO, it would
recover and become healthy again. At least, usually it recovers. If not,
and AI stays broken, then there is still the useful timeout to give up
eventually. The health check is harmfull, because AI has a certain
chance to recover.

For example, see [1] and the previous failures.

This reverts commit b1681deee4443fc2080a452f6714285ffacb42d9.

[1] https://jenkins-csb-nst-net-hw-ci.dno.corp.redhat.com/view/DPU%20Testing/job/99_E2E_Marvell_DPU_Deploy/761/