feat: Replace RHEL ISO with CentOS-Stream
An awesome feature bootc allows is that the ISO distribution doesn't really matter.
All we need is an iso that uses a kickstart and has the iscsi and other modules we
need and the ostreecontainer directive in the kickstart. Therefore we can easily deploy RHEL on the DPU using RHEL and
Centos boot iso. Currently fedora CoreOS doesn't seem to have the iscsi
modules by default. Switching to CentOS-Stream for the ISO allows users
without internal Red Hat access to build these images.