marvell: boot live ISO for accessing worker node
We use pxeboot to install RHEL on the DPU. For that, we need to SSH into
the host (that has the DPU inside) and run a podman command. That
requires that the host is up and accessible via SSH.

Also, CDA is not told by configuration which kind of DPU the host has
inside. Instead, to detect that it should install a Marvell DPU, it
needs to SSH into the host and check lspci.

If the host is not accessible, then previously installation would fail.

Well, the intended approach would have been to install the OCP cluster
(including the worker) first. Then the worker node would be accessible
via SSH, and all would be good. However, that is not done.

Hence, if the machine is not accessible, fallback to boot a CoreOS Live
iso.

https://issues.redhat.com/browse/MDC-108
https://issues.redhat.com/browse/IIC-806