ipu: don't crash during IPUBMC.ensure_started()
IPUBMC.ensure_started() called

  self._create_imc_rsh().wait_ping()

But _create_imc_rsh() tries to SSH into the host, with a timeout of one
hour. On failure (and after one hour), it fails with an exception that
was not expected by the caller and lead to a crash.

Instead, use a timeout of "only" 5 minutes, and hide the exception.
There is no need for Host.wait_ping(), because Host.ssh_connect()
already does a ping internally.

This works around the most glaring problems with this. On Marvell
cluster, the "is_ipu()" check would wait to start the machine, which
then fails to be accessible via SSH. More patches will follow to solve
this better.