bmc: let detect_dpu() return a BaseBMC instance
This is a small stap towards unifying the object hierarchy. detect_dpu()
is now basically a factory method for the BaseBMC instance, and does not
merely return a flavor string. That is useful, because detect_dpu()
already needs to create those BMC instances to do the detection. And
later on, we will need them again.

This will allow us further unification next.