bmc: let DPU detection avoid rebooting host
As it was previously, we would first try to detect an IPU and then try
to detect a Marvell DPU.

But if the BMC does not appear to be of the expected kind, previously
the detection would try harder to first boot the system (or boot
CoreOS). If you are on a Marvell system, then the IPU detection would
pointlessly try to boot the IPU system. This is despite the Marvell
system possibly already being up.

What is much better, is if we first try to detect the DPU kind without
resorting to a more expensive boot. If that succeeds for any DPU kind,
we are good. Only if we still fail to detect the system, resort to a
much more expensive reboot.

Note that we detect the DPU multiple times. So even if the first time
the Marvell BMC is down and we need the expensive approach, the second
time it will be up and we don't have the is_ipu() check first trying to
boot another system.