common: avoid dumping huge amount of logs from podman
As we can see in `podman inspect assisted-installer-service`, the
LogConfig.Type is "journald". This means, in the journal we have huge
amounts of logs from the service that are not relevant.

We then call `podman logs --events-backend=file
assisted-installer-service` and collect gigabytes of logs in our python
application. Also, we log them at DEBUG level, which fills up the
Jenkins storage.

Instead, restrict the logs that we fetch to the once since we started
the application.